rakuten-trading-bot/requirements.txt
fastapi
uvicorn[standard]
playwright
pandas
pandas_ta
streamlit
discord.py
python-dotenv
requests
